+++ packages/react-components/src/components/CardDoDont/index.ts
@@ -1,0 +1,2 @@
+export { CardDoDont } from './CardDoDont';
+export type { CardDoDontItem, CardDoDontProps } from './CardDoDont';